--- src/data-objects/data-object.ts
@@ -1,8 +1,0 @@
-import { DB } from "@/db/types";
-export abstract class DataObject {
-  protected db: DB;
-
-  constructor(db: DB) {
-    this.db = db;
-  }
-}
