--- src/data-objects/users/super-admin.ts
@@ -1,3 +1,0 @@
-import { User } from "./user";
-
-export class SuperAdmin extends User {}
